pedazo de mierda de texturas

+++ vista/src/vista/figura/FiguraView.h
@@ -23,7 +23,7 @@ private:
 protected:
 	Transformacion tl;
 public:
-	FiguraView(int x, int y, int w, int h, SDL_Texture* textura);
+	FiguraView(float x, float y, float w, float h, SDL_Texture* textura);
 	virtual ~FiguraView();
 	void dibujarse(SDL_Renderer*);
 	void dibujarse(SDL_Renderer*, SDL_Rect&);
